input: misc: add pwrkey-sync software driver
This software driver listens to power button events and performs sync when
it's held down for long.

This is to minimize data corruption when the system is not responding and
the user wants to initiate a forced reboot.

Additionally, it can forcefully trigger a kernel panic to help debugging a
kernel issue further by storing its log to something like pstore before the
hardware resets itself.

[Tuan Anh: Enable it in defconfig]
Change-Id: I87148b9e8daa45baf0557a127da4f236ed2b3486
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>